backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
pydantic>=2.10.0
email-validator>=2.2.0
python-dotenv>=1.0.1
sqlalchemy>=2.0.30
psycopg[binary]>=3.1.19
